requirements.txt
openai
flask
python-dotenv
streamlit
requests